+++ source/games/duke/src/player_w.cpp
@@ -49,7 +49,7 @@ int operateTripbomb(int snum);
 //
 //---------------------------------------------------------------------------
 
-void DoFire(struct player_struct* p, short snum)
+void DoFire(struct player_struct* p, int snum)
 {
 	int i;
 
@@ -106,7 +106,7 @@ void DoFire(struct player_struct* p, short snum)
 //
 //---------------------------------------------------------------------------
 
-void DoSpawn(struct player_struct *p, short snum)
+void DoSpawn(struct player_struct *p, int snum)
 {
 	if(!aplWeaponSpawn[p->curr_weapon][snum])
 		return;
